requirements.txt
# Core
django==4.2.13
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==22.0.0

# Async & Task Queue
celery==5.3.6
redis==5.0.4
django-redis==5.4.0
kombu==5.3.4

# Database & Time-Series
clickhouse-driver==0.2.8
django-clickhouse-backend==1.3.0

# Data / ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
transformers==4.41.1
torch==2.3.0
bertopic==0.16.0
keybert==0.8.4
sentence-transformers==3.0.1

# Scraping / Ingestion
httpx==0.27.0
beautifulsoup4==4.12.3

# Caching & Utilities
hiredis==2.3.2
python-dateutil==2.9.0
pytz==2024.1
sentry-sdk==2.4.0
structlog==24.2.0
mcp>=1.0.0

# Dev / Test
pytest==8.2.2
pytest-django==4.8.0
pytest-asyncio==0.23.7
factory-boy==3.3.0
